+++ src/main/java/app/snapshot/qure/facilities/repository/IFacilitySummaryRepository.java
@@ -4,6 +4,6 @@ import app.snapshot.qure.facilities.dto.FacilitySummaryDto;
 import org.apache.ibatis.annotations.Param;
 
 public interface IFacilitySummaryRepository {
-    FacilitySummaryDto findSummaryByFacilityId(@Param("facilityId") int facilityId);
+    FacilitySummaryDto findSummaryByFacilityId(@Param("facilityId") Long facilityId);
 
 }
