+++ src/main/java/app/snapshot/qure/facilities/service/FacilitiesService.java
@@ -29,69 +29,70 @@ public class FacilitiesService implements IFacilitiesService {
     private String publicDomain;
 
 
-    public FacilityTagDto findActiveTagByFacilityId(int facilityId) {
+    public FacilityTagDto findActiveTagByFacilityId(Long facilityId) {
         return facilityTagRepository.findActiveTagByFacilityId(facilityId);
     }
 
-    public FacilityTagDto findTagById(int tagId) {
+    public FacilityTagDto findTagById(Long tagId) {
         return facilityTagRepository.findById(tagId);
     }
 
     @Override
-    public List<FacilitiesDto> getFacilitiesListByManager(int managerId) {
+    public List<FacilitiesDto> getFacilitiesListByManager(Long managerId) {
         return facilitiesRepository.getFacilitiesListByManager(managerId);
     }
 
     @Override
-    public List<FacilitiesDto> searchFacilitiesByManager(int managerId, String q) {
+    public List<FacilitiesDto> searchFacilitiesByManager(Long managerId, String q) {
         return facilitiesRepository.searchFacilitiesByManager(managerId, q);
     }
 
     @Override
-    public FacilitiesDto getFacilitiesInfoByManager(int facilityId, int managerId) {
+    public FacilitiesDto getFacilitiesInfoByManager(Long facilityId, Long managerId) {
         return null;
     }
 
     @Override
-    public FacilitiesDto findByfacilityIdAndManager(int facilityId, int managerId) {
+    public FacilitiesDto findByfacilityIdAndManager(Long facilityId, Long managerId) {
         return facilitiesRepository.getFacilitiesInfoByManager(facilityId, managerId);
     }
 
     @Override
-    public int updateFacilitiesByManager(FacilitiesDto facilities, int managerId) {
+    public int updateFacilitiesByManager(FacilitiesDto facilities, Long managerId) {
         return facilitiesRepository.updateFacilitiesByManager(facilities, managerId);
     }
 
     @Override
-    public int deleteFacilitiesByManager(int facilityId, int managerId) {
+    public int deleteFacilitiesByManager(Long facilityId, Long managerId) {
         return facilitiesRepository.deleteFacilitiesByManager(facilityId, managerId);
     }
 
     @Override
-    public List<ChecklistTemplateDto> findTemplatesByFacilityIdAndManager(int facilityId, int managerId) {
+    public List<ChecklistTemplateDto> findTemplatesByFacilityIdAndManager(Long facilityId, Long managerId) {
         return facilitiesRepository.findTemplatesByFacilityIdAndManager(facilityId, managerId);
     }
 
     @Override
-    public List<InspectionDto> findInspectionByFacilityIdAndManager(int facilityId, int managerId) {
+    public List<InspectionDto> findInspectionByFacilityIdAndManager(Long facilityId, Long managerId) {
         return facilitiesRepository.findInspectionByFacilityIdAndManager(facilityId, managerId);
     }
 
     @Override
-    public List<InspectionDto> findInspectionByFacilityIdAndPeriodAndManager(int facilityId, LocalDate s, LocalDate e, int managerId) {
+    public List<InspectionDto> findInspectionByFacilityIdAndPeriodAndManager(Long facilityId, LocalDate s, LocalDate e, Long managerId) {
         return facilitiesRepository.findInspectionByFacilityIdAndPeriodAndManager(facilityId, s, e, managerId);
     }
 
     /* 등록 시에도 로그인한 관리자 ID 주입 */
+    /* 등록 시에도 로그인한 관리자 ID 주입 */
     /** 설비 등록 + QR 태그 발급 (원샷) */
     @Transactional
-    public int createFacilityWithQr(FacilitiesDto dto, int managerId) {
+    public Long createFacilityWithQr(FacilitiesDto dto, Long managerId) {
         // 1) 설비 생성 (managerId는 SQL에서 바인딩)
         facilitiesRepository.insertFacilitiesByManager(dto, managerId);
         // 이 시점에 dto.facilityId 가 selectKey로 세팅되어 있어야 함
 
         // 2) 새 TAG_ID 확보
-        int nextTagId = facilityTagRepository.getNextTagId();
+        Long nextTagId = facilityTagRepository.getNextTagId();
 
         // 3) URL 생성 후 INSERT
         FacilityTagDto tag = new FacilityTagDto();
@@ -104,17 +105,17 @@ public class FacilitiesService implements IFacilitiesService {
     }
 
     @Override
-    public void createByManager(FacilitiesDto facility, int managerId) {
+    public void createByManager(FacilitiesDto facility, Long managerId) {
         facilitiesRepository.insertFacilitiesByManager(facility, managerId);
     }
 
     // 지도 마커 찍기
-    public List<FacilityMarkerDto> findMarkersForManager(int managerId) {
+    public List<FacilityMarkerDto> findMarkersForManager(Long managerId) {
         return facilitiesRepository.findMarkersByManagerId(managerId);
     }
 
     @Transactional
-    public int backfillMissingCoordsForManager(int managerId) {
+    public int backfillMissingCoordsForManager(Long managerId) {
         List<FacilitiesDto> targets =
                 facilitiesRepository.findByManagerWithNullCoords(managerId);
